# public/swagger-ui-dist/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API for Laravel File Gallery",
    "description": "This is a API for Laravel file gallery to Amazon S3 specification.\nUser able to login and upload image or pdf format to Amazon S3. Also user able to display list of file gallery, update and delete the file. ",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:8000/api/"
    }
  ],
...
  "paths": {
    "/file-gallery": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "file gallery"
        ],
        "summary": "User file gallery",
        "description": "",
        "operationId": "userFileGallery",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            },
            "required": false,
            "description": "page number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "user_id": {
                            "type": "integer"
                          },
                          "title": {
                            "type": "string"
                          },
                          "desc": {
                            "type": "string"
                          },
                          "path_ext": {
                            "type": "string"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updated_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "deleted_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "url": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "links": {
                      "type": "object"
                    },
                    "meta": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid token key"
          }
        }
      },
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "file gallery"
        ],
        "summary": "User upload file to file gallery",
        "description": "",
        "operationId": "uploadFileGallery",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 150
                  },
                  "desc": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 255
                  },
                  "file": {
                    "description": "allowed format: jpg, jpeg, png, bmp, gif, svg, webp and pdf",
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "title",
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "desc": {
                      "type": "string"
                    },
                    "path_ext": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "deleted_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "id": {
                      "type": "integer"
                    },
                    "url": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid file / title supplied."
          }
        }
      }
    },
    "/file-gallery/{id}": {
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "file gallery"
        ],
        "summary": "User update file in file gallery",
        "description": "",
        "operationId": "updItemFileGallery",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "The file ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 150
                  },
                  "desc": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 255
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "user_id": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "desc": {
                      "type": "string"
                    },
                    "path_ext": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "deleted_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "url": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID / query / authentication"
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "file gallery"
        ],
        "summary": "User delete file in file gallery",
        "description": "",
        "operationId": "delItemFileGallery",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "The file ID"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully deleted!"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID / query / authentication"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}